feat: full-text search over captured command output
Add the SearchCommandOutput gRPC RPC (streaming OutputSearchMatch results
with highlighted snippets), the daemon client method, and the
`atuin search-output` CLI subcommand. The SearchIndex limit type is
normalized to usize.

+++ crates/atuin-daemon/src/search/index.rs
@@ -472,7 +472,7 @@ impl SearchIndex {
         &self,
         query: &str,
         filter_mode: &IndexFilterMode,
-        limit: u32,
+        limit: usize,
     ) -> impl Iterator<Item = HistoryId> {
         // Get precomputed frecency map (may be None if not yet computed)
         let frecency_map = self.frecency_map.read().clone();
